backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic>=2.10.0
anthropic==0.34.2
youtube-transcript-api>=1.0.0
yt-dlp>=2024.12.0
apscheduler>=3.10.0
requests>=2.32.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
python-dotenv==1.0.1
curl_cffi>=0.7.0
